new ordering testing in regression-util
Implemented new ordering testing in regression-util script.
Fixed an error in the Graph class (see test_template.cpp) where
isSatisfied() was checking if a dependnecy was satisfied, not if it was
marked (which is the requirement), now it checkes both (because both
must be true)
Changed from subprocess.call to subprocess.popen for output redirection
reasons.