Regression -> Integration
Regression tests are tests that are used to ensure bugs that are found
in previous versions of code are not reintroduced.
This is not what we were doing with out 'regression' testing framework.
Integration tests are tests that ensure that all components of a
software work in concert in a simulation of real use.