Supposedly adds support for different solvers
Previously was using CBC solver by default.
Configure can now be invoked to take differet solver backends (see the --with-*-solver flags)
Still defaults to CBC
BOP is not allowed, due to the methods used that it cannot handle.

There is an issue I am currently experiencing when configed with non CBC solvers where,
during runtime, the error "Linear solver not recognized" is ommited.
I believe ortools needs to be recompiled using some USE_* flag to indicate that a specific solver is to be used.